+++ app/page.tsx
@@ -7,7 +7,7 @@ import { t } from "./locales/translation";
 export default function Home() {
   const router = useRouter();
   const handleClick = () => {
-    router.push("/question");
+    router.push("/sections");
   };
   return (
     <div className="grid grid-rows-[20px_1fr_20px] items-center justify-items-center min-h-screen p-8 pb-20 gap-16 sm:p-20 font-[family-name:var(--font-geist-sans)]">
