--- app/question/components/(question-types)/(text-question)/LongText.tsx
@@ -1,31 +1,0 @@
-import { Textarea } from "@/components/ui/textarea";
-import { LongTextInputProps } from "@/app/question/types/text-types";
-import { useTextComponentChange } from "@/app/question/hooks/useTextComponentChange";
-
-const LongText = ({
-  placeholder,
-  defaultValue,
-  minLength,
-  maxLength,
-  onChange,
-}: LongTextInputProps) => {
-  const validate = (value: string) => !maxLength || value.length <= maxLength;
-
-  const { value, handleChange } = useTextComponentChange({
-    onChange,
-    validate,
-    defaultValue: defaultValue,
-  });
-
-  return (
-    <Textarea
-      placeholder={placeholder}
-      minLength={minLength}
-      maxLength={maxLength}
-      onChange={handleChange}
-      value={value}
-    />
-  );
-};
-
-export default LongText;
