+++ app/questionnaire/[id]/question/components/(question-types)/(multiple-choice-question)/index.tsx
@@ -1,0 +1,14 @@
+import { MultipleChoiceSubType } from "@/app/questionnaire/[id]/question/types/multiple-choice-types";
+import DCheckbox from "./DCheckbox";
+import { ComponentRegistry } from "@/app/questionnaire/[id]/question/types/component-types";
+import { QuestionComponent } from "../../QuestionComponent";
+
+const MULTI_CHOICE_COMPONENTS: ComponentRegistry<MultipleChoiceSubType> = {
+  [MultipleChoiceSubType.CHECKBOX]: DCheckbox,
+};
+
+export const MultipleChoiceQuestion = QuestionComponent({
+  components: MULTI_CHOICE_COMPONENTS,
+});
+
+export default MultipleChoiceQuestion;
