+++ app/questionnaire/[id]/question/components/(question-types)/(text-question)/ShortText.tsx
@@ -1,0 +1,31 @@
+import { Input } from "@/components/ui/input";
+import { ShortTextInputProps } from "@/app/questionnaire/[id]/question/types/text-types";
+import { useTextComponentChange } from "@/app/questionnaire/[id]/question/hooks/useTextComponentChange";
+
+const ShortText = ({
+  placeholder,
+  defaultValue,
+  minLength,
+  maxLength,
+  onChange,
+}: ShortTextInputProps) => {
+  const validate = (value: string) => !maxLength || value.length <= maxLength;
+
+  const { value, handleChange } = useTextComponentChange({
+    onChange,
+    validate,
+    defaultValue: defaultValue,
+  });
+
+  return (
+    <Input
+      placeholder={placeholder}
+      minLength={minLength}
+      maxLength={maxLength}
+      onChange={handleChange}
+      value={value}
+    />
+  );
+};
+
+export default ShortText;
