+++ app/questionnaire/[id]/question/components/QuestionProgress.tsx
@@ -1,0 +1,15 @@
+import { Progress } from "@/components/ui/progress";
+import useQuestionStore from "@/app/questionnaire/[id]/question/store/store";
+
+const QuestionProgress = () => {
+  const { currentQuestionIndex, totalQuestions } = useQuestionStore();
+
+  return (
+    <Progress
+      value={(currentQuestionIndex / totalQuestions) * 100}
+      className="h-6 w-3/4"
+    />
+  );
+};
+
+export default QuestionProgress;
