+++ app/questionnaire/[id]/question/hooks/useQuestion.ts
@@ -1,0 +1,111 @@
+import { useQuery, useQueryClient } from "@tanstack/react-query";
+import { getQuestions } from "../lib/getQuestions";
+import { useState, useEffect, useMemo, useCallback } from "react";
+import useQuestionStore from "@/app/questionnaire/[id]/question/store/store";
+
+const BATCH_SIZE = 2;
+const STALE_TIME = 5 * 60 * 1000;
+const CACHE_TIME = 30 * 60 * 1000;
+
+export const useQuestion = (questionnaireId: string) => {
+  const queryClient = useQueryClient();
+  const [currentIndex, setCurrentIndex] = useState(0);
+  const { setCurrentQuestionIndex, setTotalQuestions } = useQuestionStore();
+
+  const batchIndex = useMemo(
+    () => Math.floor(currentIndex / BATCH_SIZE),
+    [currentIndex]
+  );
+  const offset = useMemo(() => batchIndex * BATCH_SIZE, [batchIndex]);
+
+  const queryConfig = useMemo(
+    () => ({
+      queryKey: ["questions", batchIndex, BATCH_SIZE],
+      queryFn: () =>
+        getQuestions({ questionnaireId, offset, limit: BATCH_SIZE }),
+      keepPreviousData: true,
+      staleTime: STALE_TIME,
+      cacheTime: CACHE_TIME,
+      retry: 2,
+      onError: (error: Error) => {
+        console.error("Error fetching questions:", error);
+      },
+    }),
+    [batchIndex, offset]
+  );
+
+  const {
+    data: questionnaire,
+    isLoading,
+    isError,
+    error,
+  } = useQuery(queryConfig);
+
+  useEffect(() => {
+    setCurrentQuestionIndex(currentIndex);
+  }, [currentIndex, setCurrentQuestionIndex]);
+
+  const questions = questionnaire?.questions || [];
+  const totalQuestions = questionnaire?.total || 0;
+  const currentQuestion = questions[currentIndex % BATCH_SIZE];
+  console.log(
+    "!!!!currentQuestion",
+    currentQuestion,
+    questions,
+    totalQuestions
+  );
+
+  useEffect(() => {
+    setTotalQuestions(totalQuestions);
+  }, [totalQuestions]);
+
+  const goToNextQuestion = useCallback(() => {
+    console.log("!!!!currentIndex", currentIndex, totalQuestions);
+    if (currentIndex < totalQuestions - 1) {
+      if ((currentIndex + 1) % BATCH_SIZE === 0) {
+        const nextBatchIndex = batchIndex + 1;
+        queryClient.prefetchQuery({
+          queryKey: ["questions", nextBatchIndex, BATCH_SIZE],
+          queryFn: () =>
+            getQuestions({
+              questionnaireId,
+              offset: nextBatchIndex * BATCH_SIZE,
+              limit: BATCH_SIZE,
+            }),
+          staleTime: STALE_TIME,
+          cacheTime: CACHE_TIME,
+        });
+      }
+      setCurrentIndex((prev) => prev + 1);
+    }
+  }, [currentIndex, totalQuestions, batchIndex, queryClient]);
+
+  const goToPreviousQuestion = useCallback(() => {
+    setCurrentIndex((prev) => Math.max(0, prev - 1));
+  }, []);
+
+  return useMemo(
+    () => ({
+      currentQuestion,
+      isFirstQuestion: currentIndex === 0,
+      isLastQuestion: currentIndex === totalQuestions - 1,
+      totalQuestions,
+      isLoading,
+      isError,
+      error,
+      goToNextQuestion,
+      goToPreviousQuestion,
+      currentIndex,
+    }),
+    [
+      currentQuestion,
+      currentIndex,
+      totalQuestions,
+      isLoading,
+      isError,
+      error,
+      goToNextQuestion,
+      goToPreviousQuestion,
+    ]
+  );
+};
