+++ app/questionnaire/[id]/question/lib/getAllResponses.ts
@@ -1,0 +1,9 @@
+import { API_URL } from "@/app/config/apiUrls";
+
+export const getAllResponses = async () => {
+  const response = await fetch(`${API_URL.ALL_RESPONSES}`);
+  if (!response.ok) {
+    throw new Error("Failed to get response");
+  }
+  return response.json();
+};
