+++ app/questionnaire/[id]/question/lib/getResponse.ts
@@ -1,0 +1,12 @@
+import { API_URL } from "@/app/config/apiUrls";
+
+export const getResponse = async (questionId: string) => {
+  if (!questionId) {
+    throw new Error("Question ID is required");
+  }
+  const response = await fetch(`${API_URL.RESPONSE}${questionId}`);
+  if (!response.ok) {
+    throw new Error("Failed to get response");
+  }
+  return response.json();
+};
